refactor: move EC2 instance resource to new structure
